Implement interfaces for federationsender storage
